fix(web): address analytics review feedback
